fix divide precision

+++ nuxeo-admin-center/nuxeo-admin-center/src/main/java/org/nuxeo/ecm/admin/repo/RepoStatInfo.java
@@ -146,10 +146,11 @@ public class RepoStatInfo {
         if (isVersion) {
             versions += 1;
         }
-	long t2 = System.currentTimeMillis();
-	long delta = t2 - t1;
-	if (delta == 0) delta=1;
-        speed = 1000 * getTotalNbDocs() / delta;
+        long t2 = System.currentTimeMillis();
+        long delta = t2 - t1;
+        if (delta == 0)
+            delta = 1;
+        speed = 1000 * getTotalNbDocs() / (float) delta;
     }
 
     public synchronized void addBlob(long size, Path path) {
